chore: reorganize infrastructure modules
