+++ src/Infrastructure/Categories/CategoryConfigurations.cs
@@ -1,0 +1,34 @@
+using Domain.CategoryAggregate;
+using Domain.CategoryAggregate.ValueObjects;
+
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace Infrastructure.Categories;
+
+/// <summary>
+/// Configures the tables for the <see cref="Category"/> aggregate.
+/// </summary>
+public class CategoryConfigurations : IEntityTypeConfiguration<Category>
+{
+    /// <inheritdoc/>
+    public void Configure(EntityTypeBuilder<Category> builder)
+    {
+        builder.ToTable("categories");
+
+        builder.HasKey(c => c.Id);
+
+        builder
+            .Property(c => c.Id)
+            .HasConversion(id => id.Value, value => CategoryId.Create(value))
+            .ValueGeneratedOnAdd()
+            .IsRequired();
+
+        builder
+            .Property(c => c.Name)
+            .HasMaxLength(120)
+            .IsRequired();
+
+        builder.HasIndex(c => c.Name).IsUnique();
+    }
+}
